refactor(shuttle): remove reference
