+++ src/hooks/useWeb3Network.js
@@ -11,7 +11,7 @@ import {injected} from '../utils/web3'
 import {TypeConnectWallet} from '../constants/index'
 import {isMobile} from 'react-device-detect'
 import {getContract} from '../utils/web3'
-import {ERC20_ABI} from '../abi/'
+import {ERC20_ABI} from '../abi'
 
 /**
  * doc: https://github.com/NoahZinsmeister/web3-react/tree/v6/docs#useweb3react
